Add files via upload
Commented out some stuff since I don't have GoogleAPI set up and was not able to test. I hope you guys can get this working, but it does the basic stuff.

Remember to un-comment the first line in the file, and then I put a couple comments with directions on where objects go.